+++ src/com/nvlad/yii2support/migrations/commands/CommandUpDownRedoBase.java
@@ -1,5 +1,10 @@
 package com.nvlad.yii2support.migrations.commands;
 
+import com.intellij.database.dataSource.DataSourceUiUtil;
+import com.intellij.database.dataSource.LocalDataSource;
+import com.intellij.database.psi.DbDataSource;
+import com.intellij.database.psi.DbPsiFacade;
+import com.intellij.database.util.DbImplUtil;
 import com.intellij.openapi.project.Project;
 import com.nvlad.yii2support.migrations.entities.Migration;
 import com.nvlad.yii2support.migrations.entities.MigrationStatus;
@@ -7,7 +12,6 @@ import org.jetbrains.annotations.NotNull;
 
 import javax.swing.*;
 import javax.swing.tree.DefaultMutableTreeNode;
-import javax.swing.tree.DefaultTreeModel;
 import java.time.Duration;
 import java.util.Enumeration;
 import java.util.HashMap;
@@ -91,6 +95,18 @@ abstract class CommandUpDownRedoBase extends CommandBase {
         }
     }
 
+    void syncDataSources() {
+        DbPsiFacade facade = DbPsiFacade.getInstance(myProject);
+        for (DbDataSource dataSource : facade.getDataSources()) {
+            if (dataSource.getDelegate() instanceof LocalDataSource) {
+                if (DbImplUtil.isConnected(dataSource)) {
+                    LocalDataSource localDataSource = (LocalDataSource) dataSource.getDelegate();
+                    DataSourceUiUtil.performAutoSyncTask(myProject, localDataSource);
+                }
+            }
+        }
+    }
+
     private Migration findMigration(String name) {
         for (Migration migration : myMigrations) {
             if (migration.name.equals(name)) {
@@ -101,14 +117,7 @@ abstract class CommandUpDownRedoBase extends CommandBase {
         return null;
     }
 
-    private void repaintMigrationNode(Migration migration) {
-        DefaultMutableTreeNode treeNode = findTreeNode(migration);
-        if (treeNode != null) {
-            ((DefaultTreeModel) ((JTree) myComponent).getModel()).nodeChanged(treeNode);
-        }
-    }
-
-    private DefaultMutableTreeNode findTreeNode(Migration migration) {
+    DefaultMutableTreeNode findTreeNode(Migration migration) {
         if (myComponent instanceof JTree) {
             if (treeNodeMap == null) {
                 JTree tree = (JTree) myComponent;