+++ src/com/nvlad/yii2support/views/inspections/MissedViewInspection.java
@@ -1,8 +1,8 @@
 package com.nvlad.yii2support.views.inspections;
 
-import com.google.common.io.Files;
 import com.intellij.codeInspection.ProblemsHolder;
 import com.intellij.openapi.project.Project;
+import com.intellij.openapi.util.io.FileUtilRt;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
@@ -55,7 +55,7 @@ final public class MissedViewInspection extends PhpInspection {
                         }
 
                         String key = resolve.key;
-                        if (Files.getFileExtension(key).isEmpty()) {
+                        if (FileUtilRt.getExtension(key).isEmpty()) {
                             key = key + '.' + Yii2SupportSettings.getInstance(reference.getProject()).defaultViewExtension;
                         }